Split implementation of NBGL Layout layer between Nano & Wallet
